updated to include auth header

+++ src/api/apiService.js
@@ -1,10 +1,14 @@
 const API_BASE_URL = 'https://goldfish-app-9nyhd.ondigitalocean.app/api';
 
-
 const apiService = {
     async get(endpoint) {
         try {
-            const response = await fetch(`${API_BASE_URL}/${endpoint}`);
+            const token = localStorage.getItem('jwt');
+            const response = await fetch(`${API_BASE_URL}/${endpoint}`, {
+                headers: {
+                    'Authorization': `${token}`,
+                },
+            });
 
             if (response.ok) {
                 const data = await response.json();
@@ -19,11 +23,17 @@ const apiService = {
     },
 
     async post(endpoint, data) {
+        const token = localStorage.getItem('token');
+        if (!token) {
+            window.location.href = '/login'; // redirect to login page
+        }
+
         try {
             const response = await fetch(`${API_BASE_URL}/${endpoint}`, {
                 method: 'POST',
                 headers: {
                     'Content-Type': 'application/json',
+                    Authorization: `Bearer ${token}`,
                 },
                 body: JSON.stringify(data),
             });
@@ -42,8 +52,15 @@ const apiService = {
 
     async delete(endpoint, id) {
         try {
+            const token = localStorage.getItem('jwt');
+            if (!token) {
+                window.location.href = '/login'; // redirect to login page
+            }
             const response = await fetch(`${API_BASE_URL}/${endpoint}/${id}`, {
                 method: 'DELETE',
+                headers: {
+                    'Authorization': `${token}`,
+                },
             });
 
             if (response.ok) {
